requirements.txt
streamlit==1.32.0
google-generativeai==0.5.4
supabase==2.4.0
python-dotenv==1.0.0
pdfplumber==0.11.0
reportlab==4.1.0
python-docx==1.1.0
pandas==2.2.1
Pillow==10.2.0
bcrypt==4.1.2
